Move TestByteCodecTarget to new codec_test.go, add marshaling to test
